requirements.txt
 # Core dependencies for the application
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.25.0
psycopg2-binary>=2.9.7
# AWS logging
boto3>=1.28.0
watchtower>=1.7.3
